[MPT-123] Extracting to new file
