+++ packages/GIS-ly/ui/src/model/GIS/GeoJSON/Feature.ts
@@ -4,11 +4,10 @@ import {
   Geometry as SerialGeometry
 } from 'geojson';
 
-import { ICloneable, IEquatable } from "../../../../../../common/interfaces";
-
 import { BoundingBox } from "./BoundingBox";
 import { IGeometry } from "./Geometries/Geometry";
 import { GeoJson, GeoJsonProperties } from "./GeoJson";
+import { FeatureProperty, IFeatureProperty } from './FeatureProperty';
 import { BBoxState, GeoJsonTypes } from "./enums";
 import { GeometryBuilder, Point } from './Geometries';
 
@@ -16,66 +15,6 @@ export type FeatureOptions = {
   properties?: IFeatureProperty,
   id?: string,
   bbox?: BoundingBox
-}
-
-export type FeaturePropertyProperties = { [name: string]: any; }
-
-/**
- * Properties of features in GeoJSON files have not specified shape.
- *
- * Deriving from this base interface at least ensures that they can be manipulated consistently.
- *
- * @export
- * @interface IFeatureProperty
- * @extends {IEquatable<IFeatureProperty>}
- * @extends {ICloneable<IFeatureProperty>}
- */
-export interface IFeatureProperty
-  extends
-  FeaturePropertyProperties,
-  IEquatable<FeaturePropertyProperties>, ICloneable<FeatureProperty> {
-}
-
-export class FeatureProperty implements IFeatureProperty {
-  protected constructor() { }
-
-  equals(item: FeaturePropertyProperties): boolean {
-    const keys = Object.keys(this);
-    const itemKeys = Object.keys(item);
-    if (keys.length !== itemKeys.length) {
-      return false;
-    }
-
-    for (let i = 0; i < keys.length; i++) {
-      if (item[keys[i]] !== this[keys[i]]) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  clone(): FeatureProperty {
-    const featureProperty = new FeatureProperty();
-
-    const keys = Object.keys(this);
-    keys.forEach((key) => {
-      featureProperty[key] = this[key];
-    });
-
-    return featureProperty;
-  }
-
-  static fromJson(json: SerialGeoJsonProperties): FeatureProperty {
-    const featureProperty = new FeatureProperty();
-
-    const keys = Object.keys(json);
-    keys.forEach((key) => {
-      featureProperty[key] = json[key];
-    })
-
-    return featureProperty;
-  }
 }
 
 export interface FeatureProperties extends GeoJsonProperties {
